Initial couple of classes dealing with joining and identifying with an IRC network. Very rough draft; getting late so I'll sort this out tomorrow and get a clearer structure going.
The current working theory is that an IRCNetwork will consist of a name
for the network, such as "FreenodeIRC", and a list of known hosts in that
network, in a format like irc.freenode.net:6667. IRCIdentity represents a
username, nickname, and password which will be used to identify with a
server: most likely, a User class will have a number of IRCIdentity
objects, and an IRCNetwork will reference one of these as the identity in
use. Whether or not the client identifies with NickServ using the password
associated with the object will be at the client's discretion.

When I get chance, I'll add some communication functionality to
IRCNetwork and associated classes and have a working model for initial
authentication.